% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{AlterMAPE}
\alias{AlterMAPE}
\title{AlterMAPE Function}
\usage{
AlterMAPE(actual, pred, overadj = T)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}

\item{overadj}{Logical variable. Default is True. It will adjust MAPE to 100 if it is >100}
}
\value{
AlterMAPE value
}
\description{
This function allows you to calculate Alternative Mean Absolute Percent Error (AlterMAPE). This is basically error percentage from actual mean line.
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
AlterMAPE(actual = mtcars$mpg,pred=reg$fitted.values)

}
