% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{MADP}
\alias{MADP}
\title{MADP Function}
\usage{
MADP(actual, pred)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}
}
\value{
MADP value
}
\description{
This function allows you to calculate Mean Absolute Deviation Percent (MADP)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
MADP(actual = mtcars$mpg,pred=reg$fitted.values)

}
