% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{MAE}
\alias{MAE}
\title{MAE Function}
\usage{
MAE(actual, pred)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}
}
\value{
MAE value
}
\description{
This function allows you to calculate Mean Absolute Error (MAE)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
MAE(actual = mtcars$mpg,pred=reg$fitted.values)

}
