% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{MAPE}
\alias{MAPE}
\title{MAPE Function}
\usage{
MAPE(actual, pred, overadj = T)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}

\item{overadj}{Logical variable. Default is True. It will adjust MAPE to 100 if it is >100}
}
\value{
Mape value
}
\description{
This function allows you to calculate Mean Absolute Percentage Error (MAPE)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
MAPE(actual = mtcars$mpg,pred=reg$fitted.values)

}
