% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{MDA}
\alias{MDA}
\title{MDA Function}
\usage{
MDA(actual, pred)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}
}
\value{
MDA value
}
\description{
This function allows you to calculate Mean Directional Accuracy (MDA) which helps to find out how much prediction is following the trend of actual
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
MDA(actual = mtcars$mpg,pred=reg$fitted.values)

}
