% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{MedianAPE}
\alias{MedianAPE}
\title{MedianAPE Function}
\usage{
MedianAPE(actual, pred, overadj = T)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}

\item{overadj}{Logical variable. Default is True. It will adjust MAPE to 100 if it is >100}
}
\value{
MedianAPE value
}
\description{
This function allows you to calculate Median Absolute Percentage Error (MedianAPE)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
MedianAPE(actual = mtcars$mpg,pred=reg$fitted.values)

}
