% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{MSE}
\alias{MSE}
\title{MSE Function}
\usage{
MSE(actual, pred)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}
}
\value{
MSE value
}
\description{
This function allows you to calculate Mean Square Error (MSE)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
MSE(actual = mtcars$mpg,pred=reg$fitted.values)

}
