% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{NMAPE}
\alias{NMAPE}
\title{NMAPE Function}
\usage{
NMAPE(actual, pred, overadj = T)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}

\item{overadj}{Logical variable. Default is True. It will adjust MAPE to 100 if it is >100}
}
\value{
NMape value
}
\description{
This function allows you to calculate Normalized Mean Absolute Percentage Error (NMAPE), where denominator is maximum of actual and predicted
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
NMAPE(actual = mtcars$mpg,pred=reg$fitted.values)

}
