% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{R_sqr}
\alias{R_sqr}
\title{R_sqr Function}
\usage{
R_sqr(actual, pred)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}
}
\value{
R-square value
}
\description{
This function allows you to calculate R-Square value
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
R_sqr(actual = mtcars$mpg,pred=reg$fitted.values)

}
