% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{RMSE}
\alias{RMSE}
\title{RMSE Function}
\usage{
RMSE(actual, pred)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}
}
\value{
RMSE value
}
\description{
This function allows you to calculate Root Mean Square Error (RMSE)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
RMSE(actual = mtcars$mpg,pred=reg$fitted.values)

}
