% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_measures.R
\name{SMAPE}
\alias{SMAPE}
\title{SMAPE Function}
\usage{
SMAPE(actual, pred, overadj = T)
}
\arguments{
\item{actual}{Actual value}

\item{pred}{Predicted value}

\item{overadj}{Logical variable. Default is True. It will adjust MAPE to 100 if it is >100}
}
\value{
SMAPE value
}
\description{
This function allows you to calculate Symmetric Mean Absolute Percentage Error (SMAPE)
}
\examples{
data(mtcars)
reg <- lm(mpg ~ ., data = mtcars)
SMAPE(actual = mtcars$mpg,pred=reg$fitted.values)

}
